moved properties macro things into its own file, ready for entity macro
